+++ lib/local/LandmarkDetector/include/Patch_experts.h
@@ -96,10 +96,10 @@ public:
 	// The computation also requires the current landmark locations to compute response around, the PDM corresponding to the desired model, and the parameters describing its instance
 	// Also need to provide the size of the area of interest and the desired scale of analysis
 	void Response(vector<cv::Mat_<float> >& patch_expert_responses, cv::Matx22f& sim_ref_to_img, cv::Matx22f& sim_img_to_ref, const cv::Mat_<uchar>& grayscale_image, 
-							 const PDM& pdm, const cv::Vec6d& params_global, const cv::Mat_<float>& params_local, int window_size, int scale);
+							 const PDM& pdm, const cv::Vec6f& params_global, const cv::Mat_<float>& params_local, int window_size, int scale);
 
 	// Getting the best view associated with the current orientation
-	int GetViewIdx(const cv::Vec6d& params_global, int scale) const;
+	int GetViewIdx(const cv::Vec6f& params_global, int scale) const;
 
 	// The number of views at a particular scale
 	inline int nViews(size_t scale = 0) const { return (int)centers[scale].size(); };
